spring cleaning: dead code, deprecated APIs, dependency bumps
  Went through the whole codebase and threw out everything that was
  just sitting there: unused screens and helpers, commented-out old
  implementations, "// removed X" style comments, orphaned drawables,
  strings and colors nobody references anymore. Also killed a whole
  offline-dialog flow that literally could never trigger.

  While at it, moved off the deprecated Material3 bottom sheet API
  (one shared helper now instead of 20 call sites), fixed the small
  compiler warnings, and bumped basically everything: Kotlin 2.4.0,
  Compose BOM 2026.06.01, Material3 alpha23, Hilt 2.60.1, Ktor,
  JUnit and friends.

  Net result: ~3k lines lighter, zero trivial warnings, everything
  builds and tests green (debug + release/R8).

+++ app/src/main/java/com/lostf1sh/pixelplayeross/presentation/components/RoundedParallaxCarousell.kt
@@ -306,16 +306,6 @@ private fun RoundedCarousel(
             }
         }
 
-//        val clipShape = remember {
-//            object : Shape {
-//                override fun createOutline(size: Size, layoutDirection: LayoutDirection, density: Density): Outline {
-//                    return Outline.Rectangle(carouselItemInfo.maskRect) // <-- RECTANGLE: causes the flat clipping
-//                }
-//            }
-//        }
-
-        //val clipShape = rememberRoundedClipShape(carouselItemInfo, itemCornerRadius)
-
         val animatedAlpha by animateFloatAsState(
             targetValue = if (carouselStyle == CarouselStyle.ONE_PEEK && page > state.pagerState.currentPage + 1) 0f else 1f,
             animationSpec = tween(durationMillis = 200),
@@ -576,7 +566,7 @@ private class CarouselItemModifierNode(
                 }
 
                 // --- also constrain to the layer itself (safe)
-                val layerBounds = Rect(0f, 0f, size.width.toFloat(), size.height.toFloat())
+                val layerBounds = Rect(0f, 0f, size.width, size.height)
                 val maskRect = Rect(left, top, right, bottom).intersect(layerBounds)
 
                 // --- update mask info (for MaskScope, etc.)
